+++ ZendeskSell/AssociatedContacts/IAssociatedContactActions.cs
@@ -3,8 +3,8 @@ using ZendeskSell.Models;
 
 namespace ZendeskSell.AssociatedContacts {
     public interface IAssociatedContactActions {
-        Task<ZendeskSellCollectionResponse<AssociatedContactResponse>> GetAsync(int pageNumber, int numPerPage, int dealID);
-        Task<ZendeskSellObjectResponse<AssociatedContactResponse>> CreateAsync(int dealID, AssociatedContactRequest association);
-        Task<ZendeskSellDeleteResponse> DeleteAsync(int dealID, int contactID);
+        Task<ZendeskSellCollectionResponse<AssociatedContactResponse>> GetAsync(int pageNumber, int numPerPage, long dealID);
+        Task<ZendeskSellObjectResponse<AssociatedContactResponse>> CreateAsync(long dealID, AssociatedContactRequest association);
+        Task<ZendeskSellDeleteResponse> DeleteAsync(long dealID, long contactID);
     }
 }
